Correct rotation: Not ok to add an incremental rotation matrix at each time-step. Theta_rot should be used and the rotation matrix build from it at each time-step to avoid floating arithmetic error that makes the rotation matrix unorthogonal

+++ src/BeamDynLib_CBind.f90
@@ -268,7 +268,7 @@ MODULE BeamDynLib_CBind
         CALL CheckError(BD_UsrData(idxBeam), ErrStat, ErrMsg, "during restart initialization")
 
         t  = REAL(BD_UsrData(idxBeam)%t, C_DOUBLE)
-        nt = REAL(BD_UsrData(idxBeam)%nt, C_INT)
+        nt = INT(BD_UsrData(idxBeam)%nt, C_INT)
         
 
     END SUBROUTINE readRestartFile
